lambda_app/requirements.txt
boto3
botocore
PyYAML
urllib3
python-dateutil
jmespath
s3transfer
pytest
